+++ changelog.md
@@ -12,6 +12,7 @@
 - Fixed planet buttons being clickable outside of the visible scroll area.
 - Fixed solar systems in the planet screen being ordered randomly. It's now sorted alphabetically.
 - Made planet names fallback to their IDs in title case if it can't find the translations
+- Fixed Etrionic Blast Furnace Cheese (#494)
 - Added Russian translations (#509) - MaleSkam
 - Added Dutch translations (#525) - marlester-dev
 - Added Chinese translations (#523) - whyvo1